Use common code to make trusting http client in both github and montagu helpers. Update hoverfly file
